rp/requirements.txt
beautifulsoup4==4.9.1
boto3==1.15.1
botocore==1.18.1
certifi==2020.6.20
chardet==3.0.4
ChatterBot==1.0.2
chatterbot-corpus==1.2.0
click==7.1.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
joblib==0.16.0
mathparse==0.1.2
nltk==3.5
numpy==1.19.2
packaging==20.4
pi3d==2.34
Pint==0.16
PyAudio==0.2.11
pygame==1.9.6
pymongo==3.11.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.7.5
python-Levenshtein==0.12.0
pyttsx3==2.90
pytz==2020.1
PyYAML==3.13
regex==2020.7.14
requests==2.24.0
RPi.GPIO==0.7.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
svg.path==4.0.2
tqdm==4.49.0
urllib3==1.25.10
zipp==3.1.0
